fix: resolve all pre-existing test and compiler failures from RuntimeKind::Script/Wasm removal
- brassclaw_host_runtime: update all test fixtures to use RuntimeKind::Mcp (wasm/script removed)
  - host_runtime_contract.rs: process_start runtime Mcp, planner tests, service_guard tests,
    runtime_health tests, sandbox_process docker tests ignored (require Docker daemon)
  - host_runtime_services_contract.rs: add EchoMcpExecutor + with_runtime_http_egress to
    approval_resume_fixture (MCP requires network; invocation services check needs egress adapter);
    update process_backend_none → network_denied assertion; add EchoMcpExecutor to all test files
    that call SCRIPT_MANIFEST with kind=mcp
  - reborn_durable_restart_integration.rs: add EchoMcpExecutor + NoopRuntimeHttpEgress to
    base_services builder
  - reborn_invoke_vertical_slice.rs: register recording adapter under RuntimeKind::Mcp
  - runtime_policy_planner_contract.rs: rename script_runtime_requires_process_backend test →
    mcp_runtime_succeeds_without_process_backend_when_no_process_effects (MCP has no implicit
    process backend requirement, only network)
  - tool_surface_contract.rs: update allowed_runtimes FirstParty→Mcp in surface policy filter
    test; fix secret_denied_runtime_policy network_mode Deny→Brokered (MCP always requires
    network; must allow network to surface correct secret-mode denial)
  - builtin_obligation_handler_contract.rs, host_api_contract_composition.rs,
    production_trust_contract.rs, obligation_services_composition_contract.rs: wasm→mcp manifests

- brassclaw_reborn_composition/src/factory.rs: gate FilesystemCapabilityLeaseStore import under
  #[cfg(feature = "postgres")] (only used inside postgres-gated functions)
- brassclaw_reborn_composition/Cargo.toml: propagate root-llm-provider feature to
  brassclaw_product_workflow to fix feature-unification mismatch when --all activates
  brassclaw_reborn/root-llm-provider via brassclaw_reborn_cli defaults
- brassclaw_reborn_webui_ingress/tests/oidc_e2e.rs: OIDC_TEST_TOKEN_TTL_SECS u64→i64 (sign_token
  expects i64)
- brassclaw_process_sandbox/src/tests.rs: INSTALL_FAILURE_EXIT_CODE u8→i32 (DockerRunOutput and
  SandboxPhaseResult exit_code is i32)

All cargo clippy --all --all-targets -- -D warnings: CLEAN
All brassclaw_host_runtime tests: PASS (3 Docker tests ignored)

+++ crates/brassclaw_process_sandbox/src/tests.rs
@@ -1,5 +1,5 @@
 // Nonzero exit code returned by the fake Docker backend to simulate an install failure.
-const INSTALL_FAILURE_EXIT_CODE: u8 = 42;
+const INSTALL_FAILURE_EXIT_CODE: i32 = 42;
 
 use std::{
     collections::HashMap,
